Fixed signal model output
FFT outputs cosine components, added 90 deg to each freq when writing output file to change from sine to cosine. Also changed model structure. Instead of 40 sources in series, we have a libary with a model and a subcircuit file that stores the 40 harmonics from fft. It works, so that's a big plus!